functions/requirements.txt
firebase-admin==6.2.0
firebase-functions==0.1.0
numpy==1.24.3
pandas==2.0.3
scipy==1.11.4
matplotlib==3.7.2
google-cloud-storage==2.10.0